상위 브랜치와 상태 동일하게 만들기

+++ application.py
@@ -1,3 +1,4 @@
+import requests
 from flask import Flask, render_template, jsonify, request, session, redirect, url_for, make_response
 from flask_cors import CORS
 from pymongo import MongoClient
@@ -8,6 +9,7 @@ import jwt
 import hashlib
 from datetime import datetime, timedelta
 import boto3
+from apscheduler.schedulers.background import BackgroundScheduler
 
 application = Flask(__name__)
 application.config["TEMPLATES_AUTO_RELOAD"] = True
@@ -38,6 +40,18 @@ def find_pw():
     return render_template('findPw.html')
 
 
+@application.route('/search', methods=['GET'])
+def search_challenge():
+    search_receive = request.args.get('search')
+    challenges = objectIdDecoder(list(db.challenge.find({})))
+    search_ch = []
+    for challenge in challenges:
+        if search_receive in challenge["challenge_title"]:
+            search_ch.append(challenge)
+            challenge['people'] = len(list(db.join.distinct("join_user", {"join_challenge": challenge["_id"]})))
+    return render_template('index.html', challenges=search_ch)
+
+
 @application.route('/user', methods=['GET'])
 def user():
     token_receive = request.cookies.get(TOKEN_NAME)
@@ -47,11 +61,23 @@ def user():
 
         join_challenge_id_list = list(db.join.distinct("join_challenge", {'join_user': user_id}))
 
+        challenge_cnt = dict()
+        challenge_cnt['ing'] = 0
+        challenge_cnt['pause'] = 0
+        challenge_cnt['end'] = 0
+        for challenge_id in join_challenge_id_list:
+            if db.challenge.find_one({'_id': ObjectId(challenge_id)})['challenge_status'] == 1:
+                challenge_cnt['pause'] += 1
+            elif db.challenge.find_one({'_id': ObjectId(challenge_id)})['challenge_status'] == 0:
+                challenge_cnt['ing'] += 1
+            else:
+                challenge_cnt['end'] += 1
+
         challenges = []
         for challenge_id in join_challenge_id_list:
             challenges.append(db.challenge.find_one({'_id': ObjectId(challenge_id)}))
 
-        return render_template('user.html', challenges=challenges)
+        return render_template('user.html', challenges=challenges, challenge_cnt=challenge_cnt)
     except jwt.ExpiredSignatureError:
         return redirect(url_for("login", msg="로그인 시간이 만료되었습니다."))
     except jwt.exceptions.DecodeError:
@@ -86,16 +112,20 @@ def challenge_detail_page(challengeId):
         r_challenge['people'] = len(list(db.join.distinct("join_user", {"join_challenge": r_challenge['_id']})))
 
     people = len(list(db.join.distinct("join_user", {"join_challenge": challengeId})))
-
+    join = list(db.join.distinct("join_user", {"join_challenge": challengeId}))
     token_receive = request.cookies.get('fever-time')
 
     status = False
+    status_join = False
+
     try:
         payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
         status = (challenge['challenge_host'] == payload["id"])  # 내가 만든 챌리지이면 True
+        status_join = (payload["id"] in join)  # 인증한 유저 중에 내 아이디가 있으면 TRUE
     finally:
         return render_template("challenge-detail.html", challenge=challenge, people=people, status=status,
-                               categories=categories, related_challenge=related_challenge, joins=joins)
+                               categories=categories, related_challenge=related_challenge, joins=joins,
+                               status_join=status_join)
 
 
 # 준호님 code start
@@ -159,10 +189,10 @@ def check_dup():
     return jsonify({'result': 'success', 'exists': exists})
 
 
-@application.route('/check_pwd', methods=['GET'])  # 현재 비밀번호가 맞는지 확인
+@application.route('/check_pwd', methods=['POST'])  # 현재 비밀번호가 맞는지 확인
 def check_pwd():
     token_receive = request.cookies.get(TOKEN_NAME)
-    password_receive = request.args.get('pwd')
+    password_receive = request.form['pwd']
     pw_hash = hashlib.sha256(password_receive.encode('utf-8')).hexdigest()
 
     try:
@@ -197,7 +227,91 @@ def change_pwd():
     return jsonify({'result': 'success'})
 
 
+@application.route('/unregister', methods=['POST'])  # 회원탈퇴
+def unregister():
+    token_receive = request.cookies.get(TOKEN_NAME)
+    try:
+        payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+        user_id = payload['id']
+        db.users.delete_one({'user_email': user_id})
+
+
+    except jwt.ExpiredSignatureError:
+        return redirect(url_for("login", msg="로그인 시간이 만료되었습니다."))
+    except jwt.exceptions.DecodeError:
+        return redirect(url_for("login", msg="회원 정보가 존재하지 않습니다."))
+    return jsonify({"result": '회원탈퇴 되었습니다.'})
+
+
+@application.route('/oauth/callback', methods=['GET'])
+def oauth():
+    # code는 index.html에 카카오 버튼 url을 보면 알 수 있습니다. 버튼 url에 만든사람 인증id, return uri이 명시되어 있습니다.
+    # 사용자 로그인에 성공하면 로그인 한 사람의 코드를 발급해줍니다.
+    code = request.args.get("code")
+
+    # 그 코드를 이용해 서버에 토큰을 요청해야 합니다. 아래는 POST 요청을 위한 header와 body입니다.
+    client_id = '568f2b791efeffd312f12ece9bb5faea'
+    redirect_uri = 'https://fevertime.shop/oauth/callback'
+    token_url = "https://kauth.kakao.com/oauth/token"
+    token_headers = {
+        'Content-type': 'application/x-www-form-urlencoded;charset=utf-8'
+    }
+    data = {
+        'grant_type': 'authorization_code',
+        'client_id': client_id,
+        'redirect_uri': redirect_uri,
+        'code': code
+    }
+    response = requests.post(url=token_url, headers=token_headers, data=data)
+    token = response.json()
+
+    print(f'토큰 = {token}')
+    # POST 요청에 성공하면 return value를 JSON 형식으로 파싱해서 담아줍니다.
+
+    info_url = "https://kapi.kakao.com/v2/user/me"
+    info_headers = {
+        'Authorization': 'Bearer ' + token['access_token'],
+        'Content-type': 'application/x-www-form-urlencoded;charset=utf-8'
+    }
+    info_response = requests.post(url=info_url, headers=info_headers)
+    infos = info_response.json()
+    print(f'유저 정보 = {infos}')
+
+    kakao_id = infos['id']
+    kakao_name = infos['properties']['nickname']
+
+    exist = bool(db.users.find_one({'user_email': infos['id']}))
+    print(exist)
+    if exist:  # 이미 가입한 유저라면
+        payload = {
+            'id': kakao_id,
+            'exp': datetime.utcnow() + timedelta(seconds=60 * 60 * 24)  # 로그인 24시간 유지
+        }
+        token = jwt.encode(payload, SECRET_KEY, algorithm="HS256")  # 토큰을 발급하고
+        response = make_response(redirect(url_for("main_page")))  # 쿠키를 저장해줄 페이지 지정(?)
+        response.set_cookie(TOKEN_NAME, token)  # 메인페이지 기준으로 쿠키 설정(?)
+
+        return response
+    else:
+        doc = {
+            "user_email": kakao_id,
+            "user_name": kakao_name
+        }
+        db.users.insert_one(doc)
+        payload = {
+            'id': kakao_id,
+            'exp': datetime.utcnow() + timedelta(seconds=60 * 60 * 24)  # 로그인 24시간 유지
+        }
+
+        token = jwt.encode(payload, SECRET_KEY, algorithm="HS256")  # 토큰을 발급하고
+        response = make_response(redirect(url_for("main_page")))  # 쿠키를 저장해줄 페이지 지정(?)
+        response.set_cookie(TOKEN_NAME, token)  # 메인페이지 기준으로 쿠키 설정(?)
+
+        return response
+
+
 # 준호님 code end
+
 
 # 수빈님 code start
 @application.route('/challenge', methods=['POST'])
@@ -220,16 +334,17 @@ def save_challenge():
         # file_len 이 0이면 JS에서 파일을 안보낸준 것!
         # 파일을 안보내줬으면 default 파일이름을 넘겨준다.
         if file_len == 0:
-            full_file_name = "challenge.jfif"  # default 파일이름 설정
+            full_file_name = "default-challenge-img.jfif"  # default 파일이름 설정
         else:
             # 파일을 제대로 전달해줬으면 파일을 꺼내서 저장하고 파일이름을 넘겨준다.
             image_receive = request.files["image_give"]
 
             extension = image_receive.filename.split('.')[-1]
+
             today = datetime.now()
             mytime = today.strftime('%Y-%m-%d-%H-%M-%S')
 
-            filename = f'file-{mytime}'
+            filename = f'challenge-file-{mytime}'
             # save_to = f'static/assets/img/challenge/{filename}.{extension}'
             # image_receive.save(save_to)
             full_file_name = f'{filename}.{extension}'
@@ -255,7 +370,8 @@ def save_challenge():
             'challenge_endTime': period_receive.split(',')[1],
             'challenge_address': address_receive,
             'challenge_host': challenge_host,
-            'challenge_categories': categories
+            'challenge_categories': categories,
+            'challenge_status': 0,
         }
 
         db.challenge.insert_one(doc)
@@ -267,23 +383,56 @@ def save_challenge():
         return redirect(url_for("login", msg="로그인 정보가 존재하지 않습니다."))
 
 
+@application.route('/challenge', methods=['PUT'])
+def pause_challenge():
+    challengeId_receive = request.form['challengeId_give']
+    pause_receive = int(request.form['pause_give'])
+    if pause_receive == 0:
+        db.challenge.update_one({'_id': ObjectId(challengeId_receive)}, {'$set': {'challenge_status': 1}})
+        return jsonify({'result': 'success', 'msg': '챌린지가 중단 되었습니다.'})
+    else:
+        db.challenge.update_one({'_id': ObjectId(challengeId_receive)}, {'$set': {'challenge_status': 0}})
+        return jsonify({'result': 'success', 'msg': '챌린지가 활성화 되었습니다.'})
+
+
 @application.route('/challenge', methods=['DELETE'])
 def delete_challenge():
     challengeId_receive = request.form['challengeId_give']
 
     challenge_img = db.challenge.find_one({'_id': ObjectId(challengeId_receive)})['challenge_img']
-    join_img = list(db.join.find({'join_challenge': challengeId_receive})['join_img'])
+    join_list = list(db.join.find({'join_challenge': challengeId_receive}))
+
     # s3 버킷에서도 사진 삭제
     s3 = boto3.resource('s3')
     # 챌린지 이미지 삭제
-    s3.Object(os.environ["BUCKET_NAME"], challenge_img).delete()
+    if challenge_img != 'default-challenge-img.jfif':
+        s3.Object(os.environ["BUCKET_NAME"], challenge_img).delete()
     # 챌린지 인증 이미지 삭제
-    for img in join_img:
-        s3.Object(os.environ["BUCKET_NAME"], img).delete()
+    for join in join_list:
+        s3.Object(os.environ["BUCKET_NAME"], join['join_img']).delete()
 
     db.challenge.delete_one({'_id': ObjectId(challengeId_receive)})
     db.join.delete_many({'join_challenge': challengeId_receive})
     return jsonify({'result': 'success', 'msg': '챌린지 삭제 되었습니다.'})
+
+
+@application.route('/challenge/cancel', methods=['DELETE'])
+def cancel_challenge():
+    challengeId_receive = request.form['challengeId_give']
+    token_receive = request.cookies.get('fever-time')
+    payload = jwt.decode(token_receive, SECRET_KEY, algorithms=['HS256'])
+    user_receive = payload["id"]
+
+    join_list = list(db.join.find({'join_challenge': challengeId_receive, 'join_user': user_receive}))
+
+    # s3 버킷에서도 사진 삭제
+    s3 = boto3.resource('s3')
+    # 챌린지 인증 이미지 삭제
+    for join in join_list:
+        s3.Object(os.environ["BUCKET_NAME"], join['join_img']).delete()
+
+    db.join.delete_many({'join_challenge': challengeId_receive, 'join_user': user_receive})
+    return jsonify({'result': 'success', 'msg': '참가 챌린지에서 삭제 되었습니다.'})
 
 
 @application.route('/challenge/check', methods=['POST'])
@@ -294,6 +443,7 @@ def challenge_check():
 
         user_id = payload['id']
 
+        user_name = db.users.find_one({'user_email': user_id})['user_name']
         challenge_receive = request.form["challenge_give"]
         cont_receive = request.form["cont_give"]
         image_receive = request.files["img_give"]
@@ -304,7 +454,7 @@ def challenge_check():
         mytime = today.strftime("%Y-%m-%d-%H-%M-%S")
         uploadtime = today.strftime("%Y-%m-%d")
 
-        filename = f'file-{mytime}'
+        filename = f'join-file-{mytime}'
         # save_to = f'static/assets/img/join/{filename}.{extension}'
         # file.save(save_to)
         full_file_name = f'{filename}.{extension}'
@@ -326,6 +476,7 @@ def challenge_check():
             'join_challenge': challenge_receive,
             'join_date': uploadtime,
             'join_user': user_id,
+            'join_user_name': user_name,
             'join_cont': cont_receive,
             'join_img': full_file_name
         }
@@ -348,6 +499,21 @@ def objectIdDecoder(list):
         results.append(document)
     return results
 
+
+# scheduler start
+scheduler = BackgroundScheduler()
+
+
+@scheduler.scheduled_job('cron', hour='00', minute='00', id='schedule-job', timezone='Asia/Seoul')
+def challenge_scheduler():
+    today = datetime.now()
+    yesterday = today - timedelta(1)
+    date = yesterday.strftime("%Y-%m-%d")
+    db.challenge.update_many({'challenge_endTime': date}, {'$set': {'challenge_status': 2}})
+
+
+scheduler.start()
+# scheduler end
 
 # 현규님 code end
